Rename structures and functions
